fix: move examples back to top-level (#149)
These were moved as an attempt to fix the embedded examples shipped as
part of `loam init`, but the fix didn't work anyhow. There are duplicate
`core` and `status_message` examples in [crates/loam-cli/src/examples/soroban](./crates/loam-cli/src/examples/soroban)

We should fix `init` and de-duplicate the examples, yes, but in the
meantime we have a bunch of readmes that link to these examples, but all
those links are dead because the examples moved. For no good reason.